# docs/ai-prefill/openapi-preview.json
{
  "components": {
    "schemas": {
...
      "AssistanceContractResponse": {
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "envelope_schema": {
            "additionalProperties": true,
            "title": "Envelope Schema",
            "type": "object"
          },
          "profile_version": {
            "title": "Profile Version",
            "type": "string"
          },
          "validator_version": {
            "title": "Validator Version",
            "type": "string"
          }
        },
        "required": [
          "envelope_schema",
          "profile_version",
          "validator_version",
          "capabilities"
        ],
        "title": "AssistanceContractResponse",
        "type": "object"
      },
...
      "CandidateCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateEnvelope-Input",
            "description": "At most 5000 scientific objects: Compounds and their Structures each count once; locators, Lineages, members, edges, Evidence, links and Activities also count once."
          }
        },
        "required": [
          "candidate"
        ],
        "title": "CandidateCreateRequest",
        "type": "object"
      },
      "CandidateCreateResponse": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateEnvelope-Output"
          },
          "replayed": {
            "default": false,
            "title": "Replayed",
            "type": "boolean"
          }
        },
        "required": [
          "candidate"
        ],
        "title": "CandidateCreateResponse",
        "type": "object"
      },
      "CandidateEnvelope-Input": {
        "additionalProperties": false,
        "properties": {
          "candidate_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Candidate Id",
            "type": "string"
          },
          "envelope_version": {
            "const": 1,
            "title": "Envelope Version",
            "type": "integer"
          },
          "experiment_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Experiment Id",
            "type": "string"
          },
          "hashes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateHashes"
              },
              {
                "type": "null"
              }
            ]
          },
          "input_evaluation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Input Evaluation Ids",
            "type": "array"
          },
          "omissions": {
            "items": {
              "$ref": "#/components/schemas/CandidateOmission"
            },
            "title": "Omissions",
            "type": "array"
          },
          "parent_candidate_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Candidate Id"
          },
          "payload": {
            "$ref": "#/components/schemas/AiPrefillPayload-Input"
          },
          "producer": {
            "$ref": "#/components/schemas/ProducerProvenance"
          },
          "recipe": {
            "$ref": "#/components/schemas/CandidateRecipe"
          },
          "source": {
            "$ref": "#/components/schemas/SourceIdentity"
          }
        },
        "required": [
          "envelope_version",
          "candidate_id",
          "experiment_id",
          "source",
          "producer",
          "recipe",
          "payload"
        ],
        "title": "CandidateEnvelope",
        "type": "object"
      },
      "CandidateEnvelope-Output": {
        "additionalProperties": false,
        "properties": {
          "candidate_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Candidate Id",
            "type": "string"
          },
          "envelope_version": {
            "const": 1,
            "title": "Envelope Version",
            "type": "integer"
          },
          "experiment_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Experiment Id",
            "type": "string"
          },
          "hashes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateHashes"
              },
              {
                "type": "null"
              }
            ]
          },
          "input_evaluation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Input Evaluation Ids",
            "type": "array"
          },
          "omissions": {
            "items": {
              "$ref": "#/components/schemas/CandidateOmission"
            },
            "title": "Omissions",
            "type": "array"
          },
          "parent_candidate_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Candidate Id"
          },
          "payload": {
            "$ref": "#/components/schemas/AiPrefillPayload-Output"
          },
          "producer": {
            "$ref": "#/components/schemas/ProducerProvenance"
          },
          "recipe": {
            "$ref": "#/components/schemas/CandidateRecipe"
          },
          "source": {
            "$ref": "#/components/schemas/SourceIdentity"
          }
        },
        "required": [
          "envelope_version",
          "candidate_id",
          "experiment_id",
          "source",
          "producer",
          "recipe",
          "payload"
        ],
        "title": "CandidateEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreviewApplicationRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_workspace_version": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Workspace Version"
          },
          "idempotency_key": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "paper_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Id"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "validation_report": {
            "$ref": "#/components/schemas/ValidationReport"
          },
          "workspace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "validation_report",
          "idempotency_key"
        ],
        "title": "PreviewApplicationRequest",
        "type": "object"
      },
      "PreviewApplicationResponse": {
        "properties": {
          "application_id": {
            "format": "uuid",
            "title": "Application Id",
            "type": "string"
          },
          "idempotent": {
            "title": "Idempotent",
            "type": "boolean"
          },
          "receipt": {
            "additionalProperties": true,
            "title": "Receipt",
            "type": "object"
          },
          "reviewer_url": {
            "description": "Relative reviewer page URL for this application Workspace",
            "title": "Reviewer Url",
            "type": "string"
          }
        },
        "required": [
          "reviewer_url",
          "application_id",
          "idempotent",
          "receipt"
        ],
        "title": "PreviewApplicationResponse",
        "type": "object"
      },
...
      "ValidationCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "page_texts": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Page Texts",
            "type": "object"
          }
        },
        "title": "ValidationCreateRequest",
        "type": "object"
      },
      "ValidationCreateResponse": {
        "properties": {
          "replayed": {
            "default": false,
            "title": "Replayed",
            "type": "boolean"
          },
          "report": {
            "$ref": "#/components/schemas/ValidationReport"
          }
        },
        "required": [
          "report"
        ],
        "title": "ValidationCreateResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ValidationReport": {
        "additionalProperties": false,
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "candidate_sha256": {
            "title": "Candidate Sha256",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "experiment_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Experiment Id",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "payload_sha256": {
            "title": "Payload Sha256",
            "type": "string"
          },
          "profile_version": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Profile Version",
            "type": "string"
          },
          "report_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Report Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "valid",
              "needs_review",
              "invalid"
            ],
            "title": "Status",
            "type": "string"
          },
          "validator_version": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Validator Version",
            "type": "string"
          }
        },
        "required": [
          "report_id",
          "experiment_id",
          "candidate_id",
          "candidate_sha256",
          "payload_sha256",
          "profile_version",
          "validator_version",
          "status",
          "created_at"
        ],
        "title": "ValidationReport",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LeadTrace AI Prefill Preview API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/admin/ai-prefill/applications/{application_id}": {
      "get": {
        "operationId": "get_application_api_v2_admin_ai_prefill_applications__application_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "leadtrace_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadtrace Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Application",
        "tags": [
          "AI prefilling assistance"
        ]
      }
    },
    "/api/v2/admin/ai-prefill/candidates": {
      "post": {
        "operationId": "create_candidate_api_v2_admin_ai_prefill_candidates_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "leadtrace_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadtrace Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "description": "Request body exceeds 2 MiB; error code PREVIEW_REQUEST_TOO_LARGE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Candidate",
        "tags": [
          "AI prefilling assistance"
        ]
      }
    },
    "/api/v2/admin/ai-prefill/candidates/{candidate_id}": {
      "get": {
        "operationId": "get_candidate_api_v2_admin_ai_prefill_candidates__candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "leadtrace_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadtrace Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Candidate",
        "tags": [
          "AI prefilling assistance"
        ]
      }
    },
    "/api/v2/admin/ai-prefill/candidates/{candidate_id}/preview-applications": {
      "post": {
        "operationId": "apply_candidate_api_v2_admin_ai_prefill_candidates__candidate_id__preview_applications_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "leadtrace_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadtrace Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "description": "Request body exceeds 2 MiB; error code PREVIEW_REQUEST_TOO_LARGE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Candidate",
        "tags": [
          "AI prefilling assistance"
        ]
      }
    },
    "/api/v2/admin/ai-prefill/candidates/{candidate_id}/validations": {
      "post": {
        "operationId": "validate_candidate_endpoint_api_v2_admin_ai_prefill_candidates__candidate_id__validations_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "leadtrace_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadtrace Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "description": "Request body exceeds 2 MiB; error code PREVIEW_REQUEST_TOO_LARGE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Candidate Endpoint",
        "tags": [
          "AI prefilling assistance"
        ]
      }
    },
    "/api/v2/admin/ai-prefill/contracts": {
      "get": {
        "operationId": "contracts_api_v2_admin_ai_prefill_contracts_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "leadtrace_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadtrace Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistanceContractResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Contracts",
        "tags": [
          "AI prefilling assistance"
        ]
      }
    }
  }
}